+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `intersection_segment_polyline` to `compas.geometry.intersections`
 - Added `intersection_segment_polyline_xy` to `compas.geometry.intersections`
+- Added `from_sides_and_radius` to `compas.geometry.Polygon`
 
 ### Changed